move utils to find_panel
wait, something is wrong...
